move response_logger to the server package
